Move `DataSink` to `datasource` and add session crate (#15371)
* session

* clippy

* fmt

* session

* minor

* Update README.md

* doc

* fmt

* doc

---------
