+++ datafusion/catalog/Cargo.toml
@@ -36,10 +36,12 @@ async-trait = { workspace = true }
 dashmap = { workspace = true }
 datafusion-common = { workspace = true }
 datafusion-common-runtime = { workspace = true }
+datafusion-datasource = { workspace = true }
 datafusion-execution = { workspace = true }
 datafusion-expr = { workspace = true }
 datafusion-physical-expr = { workspace = true }
 datafusion-physical-plan = { workspace = true }
+datafusion-session = { workspace = true }
 datafusion-sql = { workspace = true }
 futures = { workspace = true }
 itertools = { workspace = true }
