+++ datafusion/datasource-avro/src/file_format.rs
@@ -22,26 +22,26 @@ use std::collections::HashMap;
 use std::fmt;
 use std::sync::Arc;
 
-use datafusion_common::{Result, Statistics};
-use datafusion_datasource::file_compression_type::FileCompressionType;
-use datafusion_datasource::file_format::{FileFormat, FileFormatFactory};
+use crate::avro_to_arrow::read_avro_schema_from_reader;
+use crate::source::AvroSource;
 
 use arrow::datatypes::Schema;
 use arrow::datatypes::SchemaRef;
-use async_trait::async_trait;
-use datafusion_catalog::Session;
 use datafusion_common::internal_err;
 use datafusion_common::parsers::CompressionTypeVariant;
 use datafusion_common::GetExt;
 use datafusion_common::DEFAULT_AVRO_EXTENSION;
+use datafusion_common::{Result, Statistics};
 use datafusion_datasource::file::FileSource;
+use datafusion_datasource::file_compression_type::FileCompressionType;
+use datafusion_datasource::file_format::{FileFormat, FileFormatFactory};
 use datafusion_datasource::file_scan_config::FileScanConfig;
 use datafusion_physical_expr::PhysicalExpr;
 use datafusion_physical_plan::ExecutionPlan;
-use object_store::{GetResultPayload, ObjectMeta, ObjectStore};
+use datafusion_session::Session;
 
-use crate::avro_to_arrow::read_avro_schema_from_reader;
-use crate::source::AvroSource;
+use async_trait::async_trait;
+use object_store::{GetResultPayload, ObjectMeta, ObjectStore};
 
 #[derive(Default)]
 /// Factory struct used to create [`AvroFormat`]
