+++ datafusion/datasource/src/file_sink_config.rs
@@ -15,19 +15,22 @@
 // specific language governing permissions and limitations
 // under the License.
 
+use std::sync::Arc;
+
 use crate::file_groups::FileGroup;
+use crate::sink::DataSink;
 use crate::write::demux::{start_demuxer_task, DemuxedStreamReceiver};
 use crate::ListingTableUrl;
+
 use arrow::datatypes::{DataType, SchemaRef};
-use async_trait::async_trait;
 use datafusion_common::Result;
 use datafusion_common_runtime::SpawnedTask;
 use datafusion_execution::object_store::ObjectStoreUrl;
 use datafusion_execution::{SendableRecordBatchStream, TaskContext};
 use datafusion_expr::dml::InsertOp;
-use datafusion_physical_plan::insert::DataSink;
+
+use async_trait::async_trait;
 use object_store::ObjectStore;
-use std::sync::Arc;
 
 /// General behaviors for files that do `DataSink` operations
 #[async_trait]
