+++ datafusion/datasource/src/mod.rs
@@ -39,6 +39,7 @@ pub mod file_sink_config;
 pub mod file_stream;
 pub mod memory;
 pub mod schema_adapter;
+pub mod sink;
 pub mod source;
 mod statistics;
