+++ datafusion/session/NOTICE.txt
@@ -1,0 +1,1 @@
+../../NOTICE.txt
